+++ app/controllers/creature.controller.ts
@@ -3,6 +3,7 @@ import { castleCreatures } from "../data/creatures/castle.ts";
 import { rampartCreatures } from "../data/creatures/rampart.ts";
 import { towerCreatures } from "../data/creatures/tower.ts";
 import { infernoCreatures } from "../data/creatures/inferno.ts";
+import { necropolisCreatures } from "../data/creatures/necropolis.ts";
 
 let creatures: any;
 
@@ -11,6 +12,7 @@ const factories: any = {
     'rampart': rampartCreatures,
     'tower': towerCreatures,
     'inferno': infernoCreatures,
+    'necropolis': necropolisCreatures,
 };
 
 const setFactory = (factory: string) => {
@@ -18,7 +20,8 @@ const setFactory = (factory: string) => {
 };
 
 const getDataCreatures = (factory: string) => {
-    var allCreatures = castleCreatures.concat(rampartCreatures, towerCreatures, infernoCreatures);
+    var allCreatures = castleCreatures.concat(rampartCreatures,
+      towerCreatures, infernoCreatures, necropolisCreatures);
 
     return factories[factory] || allCreatures;
 };