+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add the end point about Rampart's creatures.
 - Add the end point about Tower's creatures.
 - Add the end point about Inferno's creatures.
+- Add the end point about Necropolis's creatures.